scraper/requirements.txt
python-dotenv==1.1.0
requests==2.32.3
beautifulsoup4==4.13.4
playwright==1.52.0
prisma==0.15.0
pytest==8.1.1  # for testing
aiohttp==3.9.3  # for async HTTP requests
argparse==1.4.0
typing-extensions==4.13.2
dataclasses==0.6
